test-refactor(cargo): Made CargoLockTransformerTest a true Unit test.
